backend/requirements-auth.txt
Flask-JWT-Extended
Werkzeug
